+++ src/main/java/io/github/sweatunipd/NearYou/repository/SensorRepository.java
@@ -3,4 +3,4 @@ package io.github.sweatunipd.NearYou.repository;
 import io.github.sweatunipd.NearYou.entity.Sensor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SensorRepository extends JpaRepository<Sensor, Long> {}
+public interface SensorRepository extends JpaRepository<Sensor, Integer> {}
